When in magit-commit-mode go to that commit
Instead of going to the commit your repository is at, if you have a
magit-commit buffer open (viewing some history), go to that commit.

For example: View status of your repository in magit, then look at the
log, then move your cursor to a commit and press enter to view the
commit. At this point when you invoke github-browse-file what you expect
to happen is likely "go to this commit I'm looking at, on github".
That's what this change should do.